# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AfroMarket",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7000/afro-market/v1",
      "variables": {}
    },
    {
      "url": "https://afromarket-be.herokuapp.com/afro-market/v1",
      "variables": {}
    }
  ],
  "paths": {
...
    "/user/pass/forgot": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "forgot password",
        "operationId": "forgotpassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/forgotpasswordrequest"
                  },
                  {
                    "example": {
                      "email": "babafemi.olasunmade@zummitafrica.com"
                    }
                  }
                ]
              },
              "example": {
                "email": "babafemi.olasunmade@zummitafrica.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
    "/user/pass/reset": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "reset password",
        "operationId": "resetpassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/resetpasswordrequest"
                  },
                  {
                    "example": {
                      "newPassword": "Wolfgang9012#",
                      "otp": "179650"
                    }
                  }
                ]
              },
              "example": {
                "newPassword": "Wolfgang9012#",
                "otp": "179650"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
...
    "/product/get-all": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "get all products",
        "operationId": "getallproducts",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 10
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
...
    "/category/all": {
      "get": {
        "tags": [
          "category"
        ],
        "summary": "get all categories",
        "operationId": "getallcategories",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 10
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
...
    "/category/one/0782bc5b-4d23-4b8d-9572-052e694ad649": {
      "get": {
        "tags": [
          "category"
        ],
        "summary": "get all products in a category",
        "operationId": "getallproductsinacategory",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 1
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
...
    "/admin/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "login admin",
        "operationId": "loginadmin",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/loginadminrequest"
                  },
                  {
                    "example": {
                      "email": "babafemi.olasunmade@zummitafrica.com",
                      "password": "Wolfgang9012#"
                    }
                  }
                ]
              },
              "example": {
                "email": "babafemi.olasunmade@zummitafrica.com",
                "password": "Wolfgang9012#"
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {}
          }
        },
        "deprecated": false,
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
...
    "/admin/logout": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "log out admin",
        "operationId": "logoutadmin",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ],
        "servers": [
          {
            "url": "http://localhost:7000/afro-market/v1",
            "variables": {}
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "forgotpasswordrequest": {
        "title": "forgotpasswordrequest",
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "example": {
          "email": "babafemi.olasunmade@zummitafrica.com"
        }
      },
      "resetpasswordrequest": {
        "title": "resetpasswordrequest",
        "required": [
          "newPassword",
          "otp"
        ],
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string"
          },
          "otp": {
            "type": "string"
          }
        },
        "example": {
          "newPassword": "Wolfgang9012#",
          "otp": "179650"
        }
      },
...
      "loginadminrequest": {
        "title": "loginadminrequest",
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "babafemi.olasunmade@zummitafrica.com",
          "password": "Wolfgang9012#"
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}